Move functional dependency oracle into tests
